将get_group_card()和get_format_name()移动到Utils内
